% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctedMarkers.R
\name{correctedMarkers}
\alias{correctedMarkers}
\title{A correctedMarkers Function}
\usage{
correctedMarkers(dat, fac, con, comp)
}
\arguments{
\item{dat}{Data matrix.}

\item{fac}{Index of the group of samples.}

\item{con}{Index of correction variable.}

\item{comp}{Index (range) of measured compounds.}
}
\description{
This function allows you to perform several multiple regression on metabolomic data.
}
\examples{
correctedMarkers(iris, 5, 1, 2:4)
}
\keyword{metabolomics}
\keyword{regression}

